Don't write the final SBML document to a string.
This means that the md5 hash doesn't match the actual SBML document that's being used, but it matches the version that was loaded, so if the user re-loads that model, they'll still get the hash.  And since writing the SBML to a string is expensive, this saves a lot more time than it might occasionally cost.