Adjust published artifacts.
The wheel releases don't need the general binaries, and don't need to be buried three folders deep.

The exception is the manylinux build, which is probably the most 'universal' of the binary release sets, so continue to release those binaries, but also create a separate wheel artifact so the wheels can be found more easily.