Fix!
The trick turned out to be that on linux, loading loadBinary into jit->compiledModuleBinaryStream corrupted the compiledModuleBinaryStream.  Instead, we now convert to and from a string, and that seems to work.

At the same time, I changed the factory to return Jit* objects instead of unique_ptr objects, which also seems to work, though that might be a purely cosmetic change.