Changes test Set_Gillespie_Random_Seed to fail less.
The previous version of this test checked to see if two random walks through integers ended on different numbers, which was true every so often.  In this version, we check if a single random draw of a double is the same with two seeds, which is *much* less likely to happen.