Fix regenerateModel not setting initial values correctly.
It was getting the index wrong when there was a *dependent* *floating* species and also an *independent* *floating* species.

Greatly simplified everything:  since the SBML model now stores all initial values, we don't have to copy them over when the model is regenerated.  Also simplified setting the normal values, since that too was wrong:  now just call 'setValue' on everything, instead of trying to figure out which symbols can be set and which can't:  now, if 'setValue' works, great, and if it throws an exception, we move on with our lives.