Fix removeInitialAssignment and old tests.
When we remove the initial assignment, set the value of the remaining SBML object to the calculated initial value.

Also, the old 'reset to origin' test used to test whether the originally-loaded file was unchanged.  But that's no longer the case: when we set 'init(x)', the stored SBML changes, which means that 'resetToOrigin' will also change.

Also, update getCurrentSBML to set the initial amount or concentration according to what was set originally.