Label the matrices in the Two_Cycles rrtest, so we can re-order and pass on macs.
Also, allow the checkSVDValuesRightSingularVectors to pass by allowing absolute-value match only.  It's kind of a hack, but we don't actually use the SVD class in RoadRunner, so it's not super important.  And I *think* it's actually OK that the second row of that matrix is negative; the descriptions online I've found were kind of vague about that, though.